requirements.txt
asttokens==2.2.0
backcall==0.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
future==0.18.2
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pgeocode==0.3.0
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
pygame==2.1.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pyzmq==24.0.1
ratelim==0.1.6
requests==2.28.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tornado==6.2
traitlets==5.6.0
urllib3==1.26.13
wcwidth==0.2.5
wincertstore==0.2
